Move rsvp functions to utils.py
