+++ studygroups/tests/test_organizer_views.py
@@ -16,7 +16,7 @@ from studygroups.models import Team
 from studygroups.models import TeamMembership
 from studygroups.models import TeamInvitation
 from studygroups.models import Feedback
-from studygroups.rsvp import gen_rsvp_querystring
+from studygroups.utils import gen_rsvp_querystring
 
 import datetime
 import urllib.request, urllib.parse, urllib.error
