+++ studygroups/tests/test_tasks.py
@@ -21,8 +21,8 @@ from studygroups.models import TeamMembership
 from studygroups.models import accept_application
 from studygroups.models import create_rsvp
 from studygroups.models import generate_all_meetings
-from studygroups.rsvp import gen_rsvp_querystring
-from studygroups.rsvp import check_rsvp_signature
+from studygroups.utils import gen_rsvp_querystring
+from studygroups.utils import check_rsvp_signature
 from studygroups.utils import gen_unsubscribe_querystring
 from studygroups.utils import check_unsubscribe_signature
